requirements.txt
discord
pandas
pillow
requests
from io import BytesIO
